+++ NewSonicThing/src/entities/PlayerKnuckles.cpp
@@ -1671,7 +1671,7 @@ void PlayerKnuckles::step()
     //Animating the camera
     refreshCamera();
 
-    Global::gameMainVehicleSpeed = (int)(vel.length());
+    Global::gameMainVehicleSpeed = vel.length();
 
     //Vector3f posDiffDelta = position - prevPos;
     //printf("delta pos = %f\n\n", posDiffDelta.length()/dt);
