+++ NewSonicThing/src/entities/PlayerTails.cpp
@@ -1645,7 +1645,7 @@ void PlayerTails::step()
     //Animating the camera
     refreshCamera();
 
-    Global::gameMainVehicleSpeed = (int)(vel.length());
+    Global::gameMainVehicleSpeed = vel.length();
 
     //Vector3f posDiffDelta = position - prevPos;
     //printf("delta pos = %f\n\n", posDiffDelta.length()/dt);
